refactor: product 기본 api 시나리오 패키지 분리 (#175)
- product_api 폴더에 기본 flows 분리
- smoke 테스트 분리
