+++ k6/tests/product/smoke/category-option-smoke-test.js
@@ -5,8 +5,8 @@
  *   k6 run k6/tests/category-option/category-option-smoke-test.js --env TEST_TOKEN=
  *
  */
-import { getAuthTokenFromEnv } from '../../lib/auth-helper.js';
-import { categoryOptionFullScenario } from '../../scenarios/category-option-flows.js';
+import { getAuthTokenFromEnv } from '../../../lib/auth-helper.js';
+import { categoryOptionFullScenario } from '../../../scenarios/product_api/category-option-flows.js';
 
 export const options = {
   vus: 1,
