+++ k6/tests/product/smoke/category-smoke-test.js
@@ -5,8 +5,8 @@
  *   k6 run k6/tests/category/category-smoke-test.js --env TEST_TOKEN=
  *
  */
-import { getAuthTokenFromEnv } from '../../lib/auth-helper.js';
-import { categoryFullScenario } from '../../scenarios/category-flows.js';
+import { getAuthTokenFromEnv } from '../../../lib/auth-helper.js';
+import { categoryFullScenario } from '../../../scenarios/product_api/category-flows.js';
 
 export const options = {
   vus: 1,
