+++ k6/tests/product/smoke/product-detail-smoke-test.js
@@ -8,9 +8,9 @@
  *     --env TEST_STORE_ID=
  *
  */
-import { getAuthTokenFromEnv, getAdminToken } from '../../lib/auth-helper.js';
-import { getStoreId } from '../../lib/store-helper.js';
-import { productDetailFullScenario } from '../../scenarios/product-detail-flows.js';
+import { getAuthTokenFromEnv, getAdminToken } from '../../../lib/auth-helper.js';
+import { getStoreId } from '../../../lib/store-helper.js';
+import { productDetailFullScenario } from '../../../scenarios/product_api/product-detail-flows.js';
 
 export const options = {
   vus: 1,
